# perf: Use Cow to avoid creates String on format.