--- a/autocorrect/src/rule/rule.rs
+++ b/autocorrect/src/rule/rule.rs
@@ -1,29 +1,31 @@
+use std::borrow::Cow;
+
 use crate::config::SeverityMode;
 use crate::result::Severity;
 
 pub(crate) struct Rule {
     #[allow(dead_code)]
     pub name: String,
-    pub format_fn: fn(input: &str) -> String,
+    pub format_fn: for<'a> fn(input: &'a str) -> Cow<'a, str>,
 }
 
 #[derive(Default)]
-pub(crate) struct RuleResult {
-    pub out: String,
+pub(crate) struct RuleResult<'a> {
+    pub out: Cow<'a, str>,
     pub severity: Severity,
 }
 
-impl RuleResult {
-    pub fn new(input: &str) -> Self {
+impl<'a> RuleResult<'a> {
+    pub fn new(input: &'a str) -> Self {
         Self {
-            out: input.to_string(),
+            out: Cow::Borrowed(input),
             ..Default::default()
         }
     }
 }
 
 impl Rule {
-    pub fn new(name: &str, format: fn(input: &str) -> String) -> Self {
+    pub fn new(name: &str, format: for<'a> fn(input: &'a str) -> Cow<'a, str>) -> Self {
         Rule {
             name: name.to_string(),
             format_fn: format,
@@ -35,11 +37,10 @@ impl Rule {
             return;
         }
 
-        let new = (self.format_fn)(&result.out);
-        if result.out.ne(&new) {
+        if let Cow::Owned(new) = (self.format_fn)(&result.out) {
             result.severity = Severity::Error;
+            result.out = Cow::Owned(new);
         }
-        result.out = new;
     }
 
     pub fn lint(&self, result: &mut RuleResult) {
@@ -47,15 +48,16 @@ impl Rule {
             return;
         }
 
-        let new = (self.format_fn)(&result.out);
-        if result.out.ne(&new) && result.severity == Severity::Pass {
-            if self.severity() == SeverityMode::Warning {
-                result.severity = Severity::Warning;
-            } else {
-                result.severity = Severity::Error;
+        if let Cow::Owned(new) = (self.format_fn)(&result.out) {
+            if result.severity == Severity::Pass {
+                if self.severity() == SeverityMode::Warning {
+                    result.severity = Severity::Warning;
+                } else {
+                    result.severity = Severity::Error;
+                }
             }
+            result.out = Cow::Owned(new);
         }
-        result.out = new;
     }
 
     fn severity(&self) -> SeverityMode {
@@ -75,7 +77,7 @@ mod tests {
 
     #[test]
     fn test_rule_not_pass() {
-        let rule = Rule::new("space-word", |input| format!("{input} - foo"));
+        let rule = Rule::new("space-word", |input| Cow::Owned(format!("{input} - foo")));
         assert_eq!(rule.severity(), SeverityMode::Error);
         assert_eq!(rule.name, "space-word");
 
@@ -92,7 +94,7 @@ mod tests {
 
     #[test]
     fn test_rule_pass() {
-        let rule = Rule::new("spellcheck", |input| input.to_string());
+        let rule = Rule::new("spellcheck", |input| Cow::Borrowed(input));
 
         let mut result = RuleResult::new("test");
         rule.format(&mut result);