@huacnlee 2025-03-23
我早就想优化这部分了，里面有太多的 format 不断创建新的 String。

Cow 正好解决这个问题，每次应该只有少量的情况有修改。